+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 6.4.13(upcoming)
 ### Features
 - Added ability to boot from network for `ionoscloud_server`, `ionoscloud_vcpu_server`, `ionoscloud_cube_server`
+- Add `ionoscloud_k8s_clusters` data source
 ### Refactor
 - Remove duplicate functions for image retrieval (`checkImage`, `resolveImageName`) in `resource_volume.go`
 ### Fixes